Conflicts:
	src/Controller/ClientsController.php
	src/Controller/DocumentsController.php
	src/Template/Orders/addorder.ctp
	webroot/subpages/profile/block.php
	webroot/subpages/profile/info_order2.php

+++ src/Template/Training/edit.ctp
@@ -66,20 +66,20 @@ if (isset($quiz)){
 
 <div class="col-md-6">
     <div class="form-group">
-        <label class="control-label">Quiz Name :</label>
+        <label class="control-label">Quiz Name:</label>
             <?php if (isset($_GET["quizid"])){ echo '<input name="ID" type="hidden" value="' . $_GET["quizid"] . '">'; } ?>
             <input name="Name" class="form-control" required value="<?php if (isset($quiz)) { echo $quiz->Name; } ?>" />
     </div>
 </div>
 <div class="col-md-6">
     <div class="form-group">
-        <label class="control-label">Image :</label>
+        <label class="control-label">Image:</label>
         <input name="image" id="image" class="form-control" value="<?php if (isset($quiz)) { echo $quiz->image; } else {echo "training.png";} ?>" />
     </div>
 </div>
 <div class="col-md-6">
     <div class="form-group">
-        <label class="control-label">Attachments :</label><BR>
+        <label class="control-label">Attachments:</label><BR>
         <small>Separate your attachments with a comma</small>
         <textarea name="Attachments" class="form-control" rows="9"><?php if (isset($quiz)) { echo $quiz->Attachments; } ?></textarea>
     </div>
@@ -87,7 +87,7 @@ if (isset($quiz)){
 
 <div class="col-md-6">
     <div class="form-group">
-        <label class="control-label">Description :</label>
+        <label class="control-label">Description:</label>
         <textarea name="Description" class="form-control" rows="10" required><?php if (isset($quiz)) { echo $quiz->Description; } ?></textarea>
     </div>
 </div>
@@ -172,13 +172,16 @@ if (isset($quiz)){
                             echo '<TD>' . $question->QuestionID . '</TD>';
                             echo '<TD>' . trunc($question->Question, 75, "...") . '</TD>';
 
-                            echo '<TD><a href="editquestion?QuestionID=' . $question->QuestionID . '&new=false&quizid=' . $_GET["quizid"] . '" class="' . btnclass("EDIT") . '">Edit</a>';
-                            echo '<a href="edit?action=delete&quizid=' . $_GET["quizid"] . '&QuestionID=' . $question->QuestionID . '" class="' . btnclass("DELETE") . '"  onclick="return confirm(' . "'Are you sure you want to delete question" . $question->QuestionID  . " ?'" . ');">Delete</a></TD>';
-                            //answer($question->Answer, 0, $question->Choice0);
-                            //answer($question->Answer, 1, $question->Choice1);
-                            //answer($question->Answer, 2, $question->Choice2);
-                            //answer($question->Answer, 3, $question->Choice3);
-
+                            if (isset($_GET["answers"])){
+                                echo "<TD>" . $question->Answer . "</TD>";
+                            } else {
+                                echo '<TD><a href="editquestion?QuestionID=' . $question->QuestionID . '&new=false&quizid=' . $_GET["quizid"] . '" class="' . btnclass("EDIT") . '">Edit</a>';
+                                echo '<a href="edit?action=delete&quizid=' . $_GET["quizid"] . '&QuestionID=' . $question->QuestionID . '" class="' . btnclass("DELETE") . '"  onclick="return confirm(' . "'Are you sure you want to delete question" . $question->QuestionID . " ?'" . ');">Delete</a></TD>';
+                                //answer($question->Answer, 0, $question->Choice0);
+                                //answer($question->Answer, 1, $question->Choice1);
+                                //answer($question->Answer, 2, $question->Choice2);
+                                //answer($question->Answer, 3, $question->Choice3);
+                            }
                             echo '</TR>';
                             $index = $question->QuestionID;
                         }
@@ -188,7 +191,7 @@ if (isset($quiz)){
                 </tbody>
             </table>
 
-    <?php if ($pageit){ ?>
+    <?php if ($pageit){//don't pageit ?>
     <div class="form-actions" style="height:75px;">
                 <div class="row">
                     <div class="col-md-12" align="right">
